refactor entities and cache to model
Previous commit showed the tight coupling of
entities and cache modules - they are in fact
both physical representations of the same
database from GitHub: entities are a raw paged
response from GitHub while cache is collected
data from the responses. They are interdependent
and as such should be part of the same module to
avoid circular dependencies.

Change-Id: I54bf7ed7ac7c4d3a7da0e3f84fac844efabf5db2