users: query for organization's repositories accessible by user
... instead of any/all repositories the user has contributed to
even if they do not belong to the organization.

The important metrics in deciding which users should be members
of the organization is if they should have access to a repository
within the organization. User's activity on repositories outside
of the organization should have no bearing. Also, users should
be able to only read repositories without actually being active
contributors in them.

Due to this, the number of repositories that a user can access
can be prohibitively high for practical display and so instead
of list of repositories show their number.

Change-Id: I8221c14e5254c6b3081f707011c7574414b20aec