cmd: dry-run
Add option to execute a dry run of a command, i.e. do everything except
executing a final irreversible operation.

When executing a dry run of a command print out the operation that would
be executed instead of executing it, otherwise print out a note that the
command may be irreversible.

Change-Id: If985a330beccc0fcf4867a303fd47b855c3e9c6c