archive: (fix) archive THE least active repositories
The archive command removed repositories that were MOST active since
the given date instead of the oldest repositories.

This was because after sorting in time, GreaterThan returns the
repositories that WERE in fact active since the given date instead
of the repositories that were NOT active since the given date, which
would be the result of LessThan method.

Change-Id: Ida09e45757e83fe7d119732bacfc81b6ea0eb7e6